Single function for BBOX computations
- Consolidated the get_border function into a single function
- Added the get_bbox_border function for bbox computation
- Refactored code that uses get_border
