+++ ENIGMAsystem/SHELL/Collision_Systems/BBox/BBOXimpl.cpp
@@ -1,4 +1,5 @@
-/** Copyright (C) 2008-2013 IsmAvatar <ismavatar@gmail.com>
+/** Copyright (C) 2021 Nabeel Danish  <nabeelben@gmail.com>
+*** Copyright (C) 2008-2013 IsmAvatar <ismavatar@gmail.com>
 ***
 *** This file is a part of the ENIGMA Development Environment.
 ***
@@ -23,40 +24,12 @@
 #include "Universal_System/Object_Tiers/collisions_object.h"
 #include "Universal_System/Instances/instance_system.h" //iter
 #include "Universal_System/Instances/instance.h"
+#include "Collision_Systems/General/collisions_general.h"
 
 #include "BBOXutil.h"
 #include "BBOXimpl.h"
 #include <cmath>
 #include <floatcomp.h>
-
-static inline void get_border(int *leftv, int *rightv, int *topv, int *bottomv, int left, int top, int right, int bottom, double x, double y, double xscale, double yscale, double angle)
-{
-    if (fzero(angle))
-    {
-        const bool xsp = (xscale >= 0), ysp = (yscale >= 0);
-        const double lsc = left*xscale, rsc = (right+1)*xscale-1, tsc = top*yscale, bsc = (bottom+1)*yscale-1;
-
-        *leftv   = (xsp ? lsc : rsc) + x + .5;
-        *rightv  = (xsp ? rsc : lsc) + x + .5;
-        *topv    = (ysp ? tsc : bsc) + y + .5;
-        *bottomv = (ysp ? bsc : tsc) + y + .5;
-    }
-    else
-    {
-        const double arad = angle*(M_PI/180.0);
-        const double sina = sin(arad), cosa = cos(arad);
-        const double lsc = left*xscale, rsc = (right+1)*xscale-1, tsc = top*yscale, bsc = (bottom+1)*yscale-1;
-        const int quad = int(fmod(fmod(angle, 360) + 360, 360)/90.0);
-        const bool xsp = (xscale >= 0), ysp = (yscale >= 0),
-                   q12 = (quad == 1 || quad == 2), q23 = (quad == 2 || quad == 3),
-                   xs12 = xsp^q12, sx23 = xsp^q23, ys12 = ysp^q12, ys23 = ysp^q23;
-
-        *leftv   = cosa*(xs12 ? lsc : rsc) + sina*(ys23 ? tsc : bsc) + x + .5;
-        *rightv  = cosa*(xs12 ? rsc : lsc) + sina*(ys23 ? bsc : tsc) + x + .5;
-        *topv    = cosa*(ys12 ? tsc : bsc) - sina*(sx23 ? rsc : lsc) + y + .5;
-        *bottomv = cosa*(ys12 ? bsc : tsc) - sina*(sx23 ? lsc : rsc) + y + .5;
-    }
-}
 
 static inline int min(int x, int y) { return x<y? x : y; }
 static inline double min(double x, double y) { return x<y? x : y; }
@@ -72,12 +45,9 @@ enigma::object_collisions* const collide_inst_inst(int object, bool solid_only, 
         return NULL;
     }
 
-    const enigma::BoundingBox &box = inst1->$bbox_relative();
-    const double xscale1 = inst1->image_xscale, yscale1 = inst1->image_yscale,
-                 ia1 = inst1->image_angle;
+    // Retrieving BBOX
     int left1, top1, right1, bottom1;
-
-    get_border(&left1, &right1, &top1, &bottom1, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale1, yscale1, ia1);
+    get_bbox_border(left1, top1, right1, bottom1, inst1);
 
     for (enigma::iterator it = enigma::fetch_inst_iter_by_int(object); it; ++it)
     {
@@ -89,12 +59,8 @@ enigma::object_collisions* const collide_inst_inst(int object, bool solid_only, 
         if (inst2->sprite_index == -1 && inst2->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box2 = inst2->$bbox_relative();
-        const double x2 = inst2->x, y2 = inst2->y,
-                     xscale2 = inst2->image_xscale, yscale2 = inst2->image_yscale,
-                     ia2 = inst2->image_angle;
         int left2, top2, right2, bottom2;
-        get_border(&left2, &right2, &top2, &bottom2, box2.left(), box2.top(), box2.right(), box2.bottom(), x2, y2, xscale2, yscale2, ia2);
+        get_bbox_border(left2, top2, right2, bottom2, inst2);        
 
         if (left1 <= right2 && left2 <= right1 && top1 <= bottom2 && top2 <= bottom1)
             return inst2;
@@ -125,12 +91,8 @@ enigma::object_collisions* const collide_inst_rect(int object, bool solid_only, 
          if (inst->sprite_index == -1 && inst->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box = inst->$bbox_relative();
-        const double x = inst->x, y = inst->y,
-                     xscale = inst->image_xscale, yscale = inst->image_yscale,
-                     ia = inst->image_angle;
         int left, top, right, bottom;
-        get_border(&left, &right, &top, &bottom, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale, yscale, ia);
+        get_bbox_border(left, top, right, bottom, inst);
 
         if (left <= x2 && x1 <= right && top <= y2 && y1 <= bottom)
             return inst;
@@ -154,12 +116,8 @@ enigma::object_collisions* const collide_inst_line(int object, bool solid_only, 
         if (inst->sprite_index == -1 && inst->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box = inst->$bbox_relative();
-        const double x = inst->x, y = inst->y,
-                     xscale = inst->image_xscale, yscale = inst->image_yscale,
-                     ia = inst->image_angle;
         int left, top, right, bottom;
-        get_border(&left, &right, &top, &bottom, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale, yscale, ia);
+        get_bbox_border(left, top, right, bottom, inst);
 
         double minX = max(min(x1,x2),left);
         double maxX = min(max(x1,x2),right);
@@ -212,12 +170,8 @@ enigma::object_collisions* const collide_inst_point(int object, bool solid_only,
         if (inst->sprite_index == -1 && inst->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box = inst->$bbox_relative();
-        const double x = inst->x, y = inst->y,
-                     xscale = inst->image_xscale, yscale = inst->image_yscale,
-                     ia = inst->image_angle;
         int left, top, right, bottom;
-        get_border(&left, &right, &top, &bottom, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale, yscale, ia);
+        get_bbox_border(left, top, right, bottom, inst);
 
         if (x1 >= left && x1 <= right && y1 >= top && y1 <= bottom)
             return inst;
@@ -259,12 +213,8 @@ enigma::object_collisions* const collide_inst_ellipse(int object, bool solid_onl
         if (inst->sprite_index == -1 && inst->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box = inst->$bbox_relative();
-        const double x = inst->x, y = inst->y,
-                     xscale = inst->image_xscale, yscale = inst->image_yscale,
-                     ia = inst->image_angle;
         int left, top, right, bottom;
-        get_border(&left, &right, &top, &bottom, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale, yscale, ia);
+        get_bbox_border(left, top, right, bottom, inst);
 
         const bool intersects = line_ellipse_intersects(rx, ry, left-x1, top-y1, bottom-y1) ||
                                  line_ellipse_intersects(rx, ry, right-x1, top-y1, bottom-y1) ||
@@ -289,12 +239,8 @@ void destroy_inst_point(int object, bool solid_only, int x1, int y1)
         if (inst->sprite_index == -1 && inst->mask_index == -1) //no sprite/mask then no collision
             continue;
 
-        const enigma::BoundingBox &box = inst->$bbox_relative();
-        const double x = inst->x, y = inst->y,
-                     xscale = inst->image_xscale, yscale = inst->image_yscale,
-                     ia = inst->image_angle;
         int left, top, right, bottom;
-        get_border(&left, &right, &top, &bottom, box.left(), box.top(), box.right(), box.bottom(), x, y, xscale, yscale, ia);
+        get_bbox_border(left, top, right, bottom, inst);
 
         if (x1 >= left && x1 <= right && y1 >= top && y1 <= bottom)
             enigma_user::instance_destroy(inst->id);
